remove extra arguments to dev.read()
pyusb 1.0.0b1 has the 'interface' argument, but 1.0.0b2 (and later) do
not. Simply remove the extra arguments (interface and timeout).

This is not an ideal fix, but makes this call consistent with all others
that did this to avoid the pyusb arguments problem. A better solution
might be to create our own 'kbusb' class that abstracts away all of
these pyusb problems.

See #37.

Signed-off-by: Jonathon Reinhart <Jonathon.Reinhart@gmail.com>