Fix MIC calculation in kbdecrypt.
There were two problems in kbdecrypt preventing the MIC from
being properly calculated.  First, the data/MIC boundar was
not being adjusted by scapy-com when nwk_seclevel is changed.
This is because scapy-com splits data and MIC when the Packet
is created, according the 'nwk_seclevel' field in the Zigbee
NWK Security header.  However, the Zigbee specification defines
the 'nwk_seclevel' to be a system-wide property rather than being
something that is determined from each packet.  The Zigbee NWK
specification mandates that the 'nwk_seclevel' field is overwritten
in a received packet with the system-wide value before the packet
is processed for decryption/authentication.  Consequently, many
Zigbee implementations do not set nwk_seclevel in the transmitted
packets, so scapy-com splits data/MIC incorrectly.

To fix this in kbdecrypt, not only do we have to set nwk_seclevel
to the 'system-wide' value being used (which has been hard coded as
the commonly used value 5), but we need scapy-com to re-generate a
Packet based on this new encoding.  Hence this fix uses scapy 'build'
to convert the modified Packet to a byte string, then calls scapy
again to rebuild the Packet from the byte string (with or without
FCS, depending on the original packet type), resulting in scapy
correctly splitting data/MIC.

The second issue was that the authentication (MIC) needs to be
calculated over both the unencrypted Zigbee headers and the
encrypted data, but kbdecrypt was not correctly calculating the
length of the Zigbee headers (2 bytes too short), resulting in
the calculation of the authentication (MIC) being incorrect.

Both of these issues are fixed, and the correct MIC value is now
calculated.  Consequently, the 'micCheck' return value (from the
original API) has been reinstated.  However, to retain compatibility
with the existing API, the 'micCheck' value is only returned if
kbdecrypt is called with a new parameter 'doMicCheck' set to True
(this parameter is set by default to False).