Add support for CC2530/1 USB dongles for sniffing.
This adds support for packet sniffing for USB dongles based
on the TI CC2530 and CC2531 chips.

Note that the FCS received from these dongles is in TI CC24XX
format, so for wireshark to display the FCS information correctly,
the "TI CC24XX FCS format" in Preferences->Protocols->IEEE802.15.4
must be selected.